create objects directory
