+++ lib/index.js
@@ -1,4 +1,4 @@
-require('./foods/foods.scss');
-require('./foods/foods.js');
-require('./meals');
+require('./stylesheets/foods.scss');
+require('./objects/foods.js');
+require('./objects/meals');
 require('./application.scss');
